requirements.txt
asgiref==3.4.1
autopep8==1.6.0
boto3==1.19.8
botocore==1.22.8
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-model-utils==4.2.0
django-storages==1.12.3
djangorestframework==3.12.4
jmespath==0.10.0
Markdown==3.3.4
Pillow==9.5.0
psycopg2-binary==2.9.6
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
mysqlclient==2.2.0
django-import-export==3.2.0
drf_yasg==1.21.7